requirements_api.txt
# API Gateway and Integration Layer Dependencies

# FastAPI and ASGI server
fastapi>=0.104.0
uvicorn[standard]>=0.24.0

# Authentication and security
PyJWT>=2.8.0
bcrypt>=4.0.1
python-multipart>=0.0.6

# WebSocket support
websockets>=12.0

# HTTP client for webhooks
aiohttp>=3.9.0

# Email support
aiosmtplib>=3.0.0

# System monitoring
psutil>=5.9.0

# Data validation
pydantic>=2.5.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
httpx>=0.25.0

# Development
black>=23.0.0
isort>=5.12.0
flake8>=6.0.0

# Documentation
mkdocs>=1.5.0
mkdocs-material>=9.4.0